# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Rest API document \u003cstyle\u003e.models {display: none !important}\u003c/style\u003e",
        "title": "Example API",
        "termsOfService": "COJ",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/login": {
            "post": {
                "description": "เข้าสู่ระบบ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "เข้าสู่ระบบ"
                ],
                "summary": "เข้าสู่ระบบ",
                "parameters": [
                    {
                        "description": "ลงชื่อเข้าสู่ระบบ",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SwagLogin"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "รายการผู้ใช้งาน",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ผู้ใช้งาน"
                ],
                "summary": "รายการผู้ใช้งาน",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SwagGetAllUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "เพิ่มผู้ใช้งาน",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ผู้ใช้งาน"
                ],
                "summary": "เพิ่มผู้ใช้งาน",
                "parameters": [
                    {
                        "description": "เพิ่มผู้ใช้งาน",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwagUserBodyIncludePassword"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SwagCreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            }
        },
        "/users/{slug}": {
            "get": {
                "description": "ข้อมูลผู้ใช้งาน",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ผู้ใช้งาน"
                ],
                "summary": "ข้อมูลผู้ใช้งาน",
                "parameters": [
                    {
                        "type": "string",
                        "description": "slug ผู้ใช้งาน",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SwagGetUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "แก้ไขผู้ใช้งาน",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ผู้ใช้งาน"
                ],
                "summary": "แก้ไขผู้ใช้งาน",
                "parameters": [
                    {
                        "type": "string",
                        "description": "slug ผู้ใช้งาน",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "แก้ไขผู้ใช้งาน",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwagUserBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SwagUpdateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "ลบผู้ใช้งาน",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ผู้ใช้งาน"
                ],
                "summary": "ลบผู้ใช้งาน",
                "parameters": [
                    {
                        "type": "string",
                        "description": "slug ผู้ใช้งาน",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SwagDeleteBase"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            }
        },
        "/users/{slug}/password": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "เปลี่ยนรหัสผ่าน",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ผู้ใช้งาน"
                ],
                "summary": "เปลี่ยนรหัสผ่าน",
                "parameters": [
                    {
                        "type": "string",
                        "description": "slug ผู้ใช้งาน",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "เปลี่ยนรหัสผ่าน",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ChangePassword"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SwagChangePasswordResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ChangePassword": {
            "type": "object",
            "required": [
                "current_password",
                "new_password"
            ],
            "properties": {
                "current_password": {
                    "description": "รหัสผ่านปัจจุบัน",
                    "type": "string",
                    "example": "password"
                },
                "new_password": {
                    "description": "รหัสผ่านใหม่",
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "models.SwagChangePasswordResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูล",
                    "type": "object"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Change Password Successfully"
                },
                "status": {
                    "description": "HTTP Status Code",
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SwagCreateUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "users": {
                            "type": "object",
                            "$ref": "#/definitions/models.SwagUser"
                        }
                    }
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Created Successfully"
                },
                "status": {
                    "description": "HTTP Status Code",
                    "type": "integer",
                    "example": 201
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SwagDeleteBase": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูล",
                    "type": "object"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Deleted Successfully"
                },
                "status": {
                    "description": "HTTP Status Code",
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SwagError400": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูล",
                    "type": "object"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Bad Request"
                },
                "status": {
                    "description": "HTTP Status Code",
                    "type": "integer",
                    "example": 400
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.SwagError404": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูล",
                    "type": "object"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Not Found"
                },
                "status": {
                    "description": "HTTP Status Code",
                    "type": "integer",
                    "example": 404
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.SwagError500": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูล",
                    "type": "object"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Internal Server Error"
                },
                "status": {
                    "description": "HTTP Status Code",
                    "type": "integer",
                    "example": 500
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.SwagGetAllUsersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "pageMeta": {
                            "type": "object",
                            "$ref": "#/definitions/models.SwagPageMeta"
                        },
                        "users": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SwagUser"
                            }
                        }
                    }
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Data Found"
                },
                "status": {
                    "description": "HTTP Status Code",
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SwagGetUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "users": {
                            "type": "object",
                            "$ref": "#/definitions/models.SwagUser"
                        }
                    }
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Data Found"
                },
                "status": {
                    "description": "HTTP Status Code",
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SwagLogin": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/models.SwagUser"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "User logged in successfully"
                },
                "status": {
                    "description": "HTTP Status Code",
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                },
                "token": {
                    "description": "token",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsI"
                }
            }
        },
...
        "models.SwagUpdateUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "users": {
                            "type": "object",
                            "$ref": "#/definitions/models.SwagUser"
                        }
                    }
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Updated Successfully"
                },
                "status": {
                    "description": "HTTP Status Code",
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.SwagUserBody": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "รูป Avatar",
                    "type": "string",
                    "example": "user01.png"
                },
                "email": {
                    "description": "อีเมล์",
                    "type": "string",
                    "example": "user01@email.com"
                },
                "firstName": {
                    "description": "ชื่อ",
                    "type": "string",
                    "example": "john"
                },
                "lastName": {
                    "description": "นามสกุล",
                    "type": "string",
                    "example": "doe"
                },
                "slug": {
                    "description": "Slug",
                    "type": "string",
                    "example": "user01"
                },
                "username": {
                    "description": "Username",
                    "type": "string",
                    "example": "user01"
                }
            }
        },
        "models.SwagUserBodyIncludePassword": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "รูป Avatar",
                    "type": "string",
                    "example": "user01.png"
                },
                "email": {
                    "description": "อีเมล์",
                    "type": "string",
                    "example": "user01@email.com"
                },
                "firstName": {
                    "description": "ชื่อ",
                    "type": "string",
                    "example": "john"
                },
                "lastName": {
                    "description": "นามสกุล",
                    "type": "string",
                    "example": "doe"
                },
                "password": {
                    "description": "รหัสผ่าน",
                    "type": "string",
                    "example": "pass1234"
                },
                "slug": {
                    "description": "Slug",
                    "type": "string",
                    "example": "user01"
                },
                "username": {
                    "description": "Username",
                    "type": "string",
                    "example": "user01"
                }
            }
        },
        "routes.login": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}